requirements.txt
discord.py
apscheduler
pytz
python-dotenv
requests
numpy
pandas
